+++ app/Domains/Auth/Http/Controllers/Backend/User/UserSessionController.php
@@ -4,12 +4,11 @@ namespace App\Domains\Auth\Http\Controllers\Backend\User;
 
 use App\Domains\Auth\Http\Requests\Backend\User\ClearUserSessionRequest;
 use App\Domains\Auth\Models\User;
-use App\Http\Controllers\Controller;
 
 /**
  * Class UserSessionController.
  */
-class UserSessionController extends Controller
+class UserSessionController
 {
     /**
      * @param  ClearUserSessionRequest  $request