+++ app/Domains/Auth/Http/Controllers/Frontend/Auth/SocialController.php
@@ -4,13 +4,12 @@ namespace App\Domains\Auth\Http\Controllers\Frontend\Auth;
 
 use App\Domains\Auth\Events\User\UserLoggedIn;
 use App\Domains\Auth\Services\UserService;
-use App\Http\Controllers\Controller;
 use Laravel\Socialite\Facades\Socialite;
 
 /**
  * Class SocialController.
  */
-class SocialController extends Controller
+class SocialController
 {
     /**
      * @param $provider