+++ app/Domains/Auth/Http/Controllers/Frontend/Auth/TwoFactorAuthenticationController.php
@@ -2,13 +2,12 @@
 
 namespace App\Domains\Auth\Http\Controllers\Frontend\Auth;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 /**
  * Class TwoFactorAuthenticationController.
  */
-class TwoFactorAuthenticationController extends Controller
+class TwoFactorAuthenticationController
 {
     /**
      * @param  Request  $request