+++ app/Domains/Auth/Http/Controllers/Frontend/Auth/UpdatePasswordController.php
@@ -4,12 +4,11 @@ namespace App\Domains\Auth\Http\Controllers\Frontend\Auth;
 
 use App\Domains\Auth\Http\Requests\Frontend\Auth\UpdatePasswordRequest;
 use App\Domains\Auth\Services\UserService;
-use App\Http\Controllers\Controller;
 
 /**
  * Class UpdatePasswordController.
  */
-class UpdatePasswordController extends Controller
+class UpdatePasswordController
 {
     /**
      * @var UserService