+++ app/Domains/Auth/Http/Controllers/Frontend/Auth/VerificationController.php
@@ -2,14 +2,13 @@
 
 namespace App\Domains\Auth\Http\Controllers\Frontend\Auth;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\VerifiesEmails;
 use Illuminate\Http\Request;
 
 /**
  * Class VerificationController.
  */
-class VerificationController extends Controller
+class VerificationController
 {
     /*
     |--------------------------------------------------------------------------
@@ -32,17 +31,6 @@ class VerificationController extends Controller
     public function redirectPath()
     {
         return route(homeRoute());
-    }
-
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->middleware('signed')->only('verify');
-        $this->middleware('throttle:6,1')->only('verify', 'resend');
     }
 
     /**