+++ app/Http/Controllers/Frontend/User/AccountController.php
@@ -2,15 +2,14 @@
 
 namespace App\Http\Controllers\Frontend\User;
 
-use App\Http\Controllers\Controller;
-
 /**
  * Class AccountController.
  */
-class AccountController extends Controller
+class AccountController
 {
+
     /**
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function index()
     {