+++ app/Http/Controllers/LocaleController.php
@@ -5,8 +5,9 @@ namespace App\Http\Controllers;
 /**
  * Class LocaleController.
  */
-class LocaleController extends Controller
+class LocaleController
 {
+
     /**
      * @param $locale
      *